+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/CertifiedProductDAO.java
@@ -1,5 +1,6 @@
 package gov.healthit.chpl.dao;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -27,6 +28,7 @@ import gov.healthit.chpl.exception.EntityRetrievalException;
 import gov.healthit.chpl.logging.Loggable;
 import gov.healthit.chpl.scheduler.job.urlStatus.data.UrlType;
 import gov.healthit.chpl.util.AuthUtil;
+import gov.healthit.chpl.util.DateUtil;
 import gov.healthit.chpl.util.ErrorMessageUtil;
 import lombok.NoArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -312,8 +314,10 @@ public class CertifiedProductDAO extends BaseDAOImpl {
 
     @Transactional(readOnly = true)
     public List<CertifiedProductDetailsDTO> findByEdition(final String edition) {
-        Query query = entityManager.createQuery("SELECT cpd " + "FROM CertifiedProductDetailsEntity cpd "
-                + "WHERE (NOT deleted = true) " + "AND cpd.year = :edition ", CertifiedProductDetailsEntity.class);
+        Query query = entityManager.createQuery("SELECT cpd "
+                + "FROM CertifiedProductDetailsEntity cpd "
+                + "WHERE (NOT deleted = true) "
+                + "AND cpd.year = :edition ", CertifiedProductDetailsEntity.class);
         query.setParameter("edition", edition.trim());
         List<CertifiedProductDetailsEntity> entities = query.getResultList();
         List<CertifiedProductDetailsDTO> products = new ArrayList<>(entities.size());
@@ -323,6 +327,28 @@ public class CertifiedProductDAO extends BaseDAOImpl {
             products.add(product);
         }
         return products;
+    }
+
+    public List<Long> findIdsByEditionAndCreatedBeforeDate(String edition, LocalDate creationDate) {
+        Query query = entityManager.createQuery("SELECT cpd.id "
+                + "FROM CertifiedProductDetailsEntity cpd "
+                + "WHERE (NOT deleted = true) "
+                + "AND cpd.year = :edition "
+                + "AND cpd.creationDate < :creationDate", Long.class);
+        query.setParameter("edition", edition.trim());
+        query.setParameter("creationDate", new Date(DateUtil.toEpochMillis(creationDate)));
+        return query.getResultList();
+    }
+
+    @Transactional(readOnly = true)
+    public List<Long> findIdsByEdition(String edition) {
+        Query query = entityManager.createQuery("SELECT cpd.id "
+                + "FROM CertifiedProductDetailsEntity cpd "
+                + "WHERE (NOT deleted = true) "
+                + "AND cpd.year = :edition ", Long.class);
+        query.setParameter("edition", edition.trim());
+        List<Long> ids = query.getResultList();
+        return ids;
     }
 
     @Transactional(readOnly = true)