+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/CertifiedProductSearchManager.java
@@ -4,6 +4,7 @@ import java.time.LocalDate;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.util.Collection;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -20,6 +21,7 @@ import gov.healthit.chpl.domain.CertificationStatus;
 import gov.healthit.chpl.domain.CertificationStatusEvent;
 import gov.healthit.chpl.domain.compliance.DirectReview;
 import gov.healthit.chpl.domain.compliance.DirectReviewNonConformity;
+import gov.healthit.chpl.domain.concept.CertificationEditionConcept;
 import gov.healthit.chpl.domain.search.CertifiedProductBasicSearchResultLegacy;
 import gov.healthit.chpl.domain.search.CertifiedProductFlatSearchResultLegacy;
 import gov.healthit.chpl.domain.search.SearchRequestLegacy;
@@ -28,6 +30,8 @@ import gov.healthit.chpl.search.domain.CertifiedProductBasicSearchResult;
 import gov.healthit.chpl.search.domain.CertifiedProductFlatSearchResult;
 import gov.healthit.chpl.search.domain.CertifiedProductSearchResult;
 import gov.healthit.chpl.service.DirectReviewSearchService;
+import gov.healthit.chpl.service.RealWorldTestingEligibility;
+import gov.healthit.chpl.service.realworldtesting.RealWorldTestingEligiblityCachingService;
 import lombok.extern.log4j.Log4j2;
 
 @Service
@@ -35,40 +39,70 @@ import lombok.extern.log4j.Log4j2;
 public class CertifiedProductSearchManager {
     private static final String CERT_STATUS_EVENT_DATE_FORMAT = "yyyy-MM-dd";
     private CertifiedProductSearchDAO searchDao;
+    private RealWorldTestingEligiblityCachingService rwtService;
     private DirectReviewSearchService drService;
     private DateTimeFormatter dateFormatter;
 
     @Autowired
-    public CertifiedProductSearchManager(CertifiedProductSearchDAO searchDao, DirectReviewSearchService drService) {
+    public CertifiedProductSearchManager(CertifiedProductSearchDAO searchDao,
+            RealWorldTestingEligiblityCachingService rwtService,
+            DirectReviewSearchService drService) {
         this.searchDao = searchDao;
+        this.rwtService = rwtService;
         this.drService = drService;
         this.dateFormatter = DateTimeFormatter.ofPattern(CERT_STATUS_EVENT_DATE_FORMAT);
     }
 
-    @Transactional(readOnly = true)
+    @Deprecated
     @Cacheable(value = CacheNames.COLLECTIONS_LISTINGS, key = "'listings'")
     public List<CertifiedProductFlatSearchResult> getFlatListingCollection() {
         List<CertifiedProductFlatSearchResult> results = searchDao.getFlatCertifiedProducts();
-        results.stream()
-            .forEach(searchResult -> populateDirectReviewFields(searchResult));
+        LOGGER.info("Populating Direct Review fields for collections");
+        Date start = new Date();
+        results.parallelStream()
+            .forEach(searchResult -> populateDirectReviews(searchResult));
+        Date end = new Date();
+        LOGGER.info("Completed Populating Direct Review fields  for collections [ " + (end.getTime() - start.getTime()) + " ms ]");
         return results;
     }
 
-    @Transactional(readOnly = true)
     @Cacheable(value = CacheNames.COLLECTIONS_SEARCH)
     public List<CertifiedProductBasicSearchResult> getSearchListingCollection() {
         List<CertifiedProductBasicSearchResult> results = searchDao.getCertifiedProducts();
-        results.stream()
-            .forEach(searchResult -> populateDirectReviewFields(searchResult));
+        LOGGER.info("Populating Direct Review fields for search");
+        Date start = new Date();
+        results.parallelStream()
+            .forEach(searchResult -> populateDirectReviews(searchResult));
+        Date end = new Date();
+        LOGGER.info("Completed Populating Direct Review fields  for search [ " + (end.getTime() - start.getTime()) + " ms ]");
+        LOGGER.info("Populating RWT Eligibility field for search");
+        start = new Date();
+        results.parallelStream()
+            .filter(listing -> listing.getEdition().equals(CertificationEditionConcept.CERTIFICATION_EDITION_2015.getYear()))
+            .forEach(listing -> populateRwtEligibility(listing));
+        end = new Date();
+        LOGGER.info("Completed Populating RWT Eligibility field for search [ " + (end.getTime() - start.getTime()) + " ms ]");
+        LOGGER.info("Populating RWT Eligibility field for 2011/2014 listings for search");
+        start = new Date();
+        results.parallelStream()
+            .filter(listing -> !listing.getEdition().equals(CertificationEditionConcept.CERTIFICATION_EDITION_2015.getYear()))
+            .forEach(listing -> listing.setIsRwtEligible(false));
+        end = new Date();
+        LOGGER.info("Completed Populating RWT Eligibility field for 2011/2014 listings for search [ " + (end.getTime() - start.getTime()) + " ms ]");
         return results;
     }
 
-    private void populateDirectReviewFields(CertifiedProductBasicSearchResult searchResult) {
+    private void populateRwtEligibility(CertifiedProductBasicSearchResult searchResult) {
+        RealWorldTestingEligibility rwtElig = rwtService.getRwtEligibility(searchResult.getId());
+        searchResult.setIsRwtEligible(rwtElig.getEligibilityYear() != null);
+    }
+
+    private void populateDirectReviews(CertifiedProductBasicSearchResult searchResult) {
         List<CertificationStatusEvent> statusEvents = createStatusEventsFromBasicSearchResult(searchResult);
         populateDirectReviewFields(searchResult, statusEvents);
     }
 
-    private void populateDirectReviewFields(CertifiedProductFlatSearchResult searchResult) {
+    private void populateDirectReviews(CertifiedProductFlatSearchResult searchResult) {
         List<CertificationStatusEvent> statusEvents = createStatusEventsFromFlatSearchResult(searchResult);
         populateDirectReviewFields(searchResult, statusEvents);
     }