+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/domain/BasicSearchResponse.java
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.annotation.JsonView;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
+@Deprecated
 public class BasicSearchResponse implements Serializable {
     private static final long serialVersionUID = 2569559170265522788L;